pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "den-bot"
version = "0.1.0"
description = "Denjamin — a Discord bot with Claude-powered chat and a points/leaderboard system."
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.4.0",
    "anthropic>=0.42.0",
    "psycopg2-binary>=2.9.10",
    "aiosql>=9.0",
    "temporalio>=1.9.0",
    "python-dotenv>=1.0",
]

[project.scripts]
denjamin = "denbot.cli:main"
temporal-worker = "denbot.worker:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"denbot.points" = ["sql/*.sql"]
"denbot.db" = ["schema.sql"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ipython>=8.12",
    "ruff>=0.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B", "SIM"]
"**/__init__.py" = ["F401"]
